refector: UserCard 컴포넌트 분리
